CRITICAL FIX: Maintain float32 for RoPE cache (FlashInfer requirement)
Root cause identified: FlashInfer requires float32 RoPE cache, but DeepSeek
was converting to float16 (model dtype), causing dispatch failures.

Key changes:
1. **Keep RoPE cache as float32**: Don't convert cos_sin_cache to model dtype
2. **Follow Llama/Qwen pattern**: Match their successful implementation
3. **Remove dtype conversion**: Let FlashInfer handle internal type conversion

This matches the working implementation in Llama/Qwen models where:
- inv_freq.dtype is float32 from HuggingFace models
- cos_sin_cache preserves this float32 dtype
- FlashInfer expects and works with float32 RoPE cache

+++ attn_hub/mla_attn.py
@@ -46,8 +46,8 @@ def apply_decoupled_rope(query_states, key_states, position_ids, rope_dim, cos_c
         # 如果已经是合并格式
         cos_sin_cache = cos_cache
     
-    # 确保数据类型与query/key一致
-    cos_sin_cache = cos_sin_cache.to(dtype=q_rope.dtype, device=q_rope.device)
+    # 确保在正确设备上，但保持float32数据类型（FlashInfer要求）
+    cos_sin_cache = cos_sin_cache.to(device=q_rope.device, dtype=torch.float32)
     
     # FlashInfer期望完整的cos_sin_cache和position_ids，内部会处理索引
     flashinfer.rope.apply_rope_with_cos_sin_cache_inplace(
